Refactor utils so standalone.events does not import eventlet
