+++ nameko/runners.py
@@ -4,7 +4,7 @@ from contextlib import contextmanager
 from logging import getLogger
 
 from nameko.containers import get_container_cls, get_service_name
-from nameko.utils import SpawningProxy
+from nameko.utils.concurrency import SpawningProxy
 
 
 _log = getLogger(__name__)
